+++ ChangeLog.md
@@ -13,6 +13,9 @@ Change Log
   * Other small enhancements
     
     - minor optimization in MultiplexorValueCalculator
+    
+    - added maximum run time for binning-converged controller to prevent
+      excessively long runs (`tomorun` option `--control-binning-converged-max-add-run-iters`)
 
 
 ## v5.1 (2017-06-24)